+++ django_yaml_fixtures/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DjangoYAMLFixturesConfig(AppConfig):
+    name = 'django_yaml_fixtures'