+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - move `py_yaml_fixtures.factory_interface.FactoryInterface` into `py_yaml_fixtures.factories`
 - `SQLAlchemyModelFactory` must now be imported from `py_yaml_fixtures.factories.sqlalchemy`
+- rename `PY_YAML_FIXTURES_MODELS_MODULE` to `FLASK_MODELS_MODULE`
 
 ## v0.3.2 (2018/12/08)