+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.4.0 (unreleased)
 
+- add Django support
 - move `py_yaml_fixtures.factory_interface.FactoryInterface` into `py_yaml_fixtures.factories`
 - `SQLAlchemyModelFactory` must now be imported from `py_yaml_fixtures.factories.sqlalchemy`
 - rename `PY_YAML_FIXTURES_MODELS_MODULE` to `FLASK_MODELS_MODULE`